move branch validations to "validations/index.js"
