+++ src/index.js
@@ -1,5 +1,6 @@
 import minimist from 'minimist';
 import { every } from 'lodash';
+import validations from './validations';
 import version from './npm/version';
 import publish from './npm/publish';
 import release from './github/release';
@@ -11,6 +12,7 @@ import config from './config';
 const _argv = minimist(process.argv.slice(2));
 
 const defaultArgv = {
+  'no-validations': false,
   'npm-publish': true,
   'no-npm-publish': false,
   'npm-version': true,
@@ -27,9 +29,14 @@ const main = async () => {
   try {
     !config.github.token && await askForToken();
 
+    !argv['no-validations'] && await validations();
+
     argv['npm-version'] && await version();
+
     argv.changelog && await changelog();
+
     argv['gh-release'] && await release({ all: false });
+
     ((runDefault && !argv['no-npm-publish']) || argv['npm-publish']) && await publish();
 
     argv['gh-release-all'] && await release({ all: true });
