+++ src/npm/version.js
@@ -1,8 +1,7 @@
 import semver from 'semver';
-import { find, sortBy, range, some } from 'lodash';
+import { find, sortBy, range } from 'lodash';
 import {
   github,
-  getCurrentBranch,
   isVersionTag,
   getPackageJsonVersion,
   info,
@@ -14,67 +13,8 @@ import {
   rl,
   exec
 } from '../utils';
-import config from '../config';
 
 const stdio = [process.stdin, null, process.stderr];
-
-const runValidations = async () => {
-  const shouldRunValidations = some(config.publish);
-
-  if (shouldRunValidations) {
-    info('Run validations');
-
-    // ENFORCE BRANCH
-    if (config.publish.branch !== null) {
-      status.addSteps(['Validate branch']);
-
-      if (getCurrentBranch() !== config.publish.branch) {
-        throw new CustomError(`You must be on "${config.publish.branch}" branch to perform this task. Aborting.`);
-      }
-      status.doneStep(true);
-    }
-
-    // ENFORCE NO UNCOMMITTED CHANGES
-    if (config.publish.noUncommittedChanges) {
-      status.addSteps(['Validate uncommitted changes']);
-
-      if (/^([ADRM]| [ADRM])/m.test(await exec('git status --porcelain'))) {
-        throw new CustomError('You have uncommited changes in your working tree. Aborting.');
-      }
-      status.doneStep(true);
-    }
-
-    // ENFORCE NO UNTRACKED FILES
-    if (config.publish.noUntrackedFiles) {
-      status.addSteps(['Validate untracked files']);
-
-      if (/^\?\?/m.test(await exec('git status --porcelain'))) {
-        throw new CustomError('You have untracked files in your working tree. Aborting.');
-      }
-      status.doneStep(true);
-    }
-
-    // ENFORCE SYNC WITH REMOTE
-    if (config.publish.inSyncWithRemote) {
-      status.addSteps(['Validate sync with remote']);
-
-      await exec('git fetch');
-
-      const LOCAL = (await exec('git rev-parse @', { encoding: 'utf8' })).trim();
-      const REMOTE = (await exec('git rev-parse @{u}', { encoding: 'utf8' })).trim();
-      const BASE = (await exec('git merge-base @ @{u}', { encoding: 'utf8' })).trim();
-
-      if (LOCAL !== REMOTE && LOCAL === BASE) {
-        throw new CustomError('Your local branch is out-of-date. Please pull the latest remote changes. Aborting.');
-      } else if (LOCAL !== REMOTE && REMOTE === BASE) {
-        throw new CustomError('Your local branch is ahead of its remote branch. Please push your local changes. Aborting.');
-      } else if (LOCAL !== REMOTE) {
-        throw new CustomError('Your local and remote branches have diverged. Please put them in sync. Aborting.');
-      }
-      status.doneStep(true);
-    }
-  }
-};
 
 const computeRelease = async (packageJsonVersion) => {
   info('Compute release');
@@ -166,9 +106,7 @@ const version = async (releaseInfo) => {
 };
 
 export default async () => {
-  title('Increase version and publish package on npm');
-
-  await runValidations();
+  title('Increase version in "package.json"');
 
   const releaseInfo = await computeRelease(getPackageJsonVersion());
